BDM-Assets/carousel-generator/requirements.txt
pyyaml>=6.0
jinja2>=3.1
playwright>=1.40
python-dotenv>=1.0
requests>=2.31
